Pushed working sort_3, operations.
Pushed working sort_3, operations. Sort_3 works for all cases, rewrote the function, along with and added helper function to reset my nodes each time it after it exits an if
statement. And made sure that the operations worked, like rotate and rrotate, isolating the first or last node to bring to the front of back of the list depending on the
rotation. Updated .h file."