23 lines for sort_3 func. Added a push func.
Managed to shorten the length of the sort_3 function from 40 to 23 lines. More efficient too with a t_value struct holding the values of 5 digits. Made a Push function from
stack_a to stack_b by allowing a temp var to hold the node to push, then set the stack_a->top to the next node, then push the node from the temp to stack_b->top. (Used for
pushing min value in stack a to stack b) Added stack_b_utils for initialising stack_b and its nodes, but not efficient, might tweak soon.